refactor : position 패키지 분리
