+++ src/main/java/chess/domain/board/ChessBoard.java
@@ -2,7 +2,7 @@ package chess.domain.board;
 
 import chess.domain.chess.ChessGame;
 import chess.domain.piece.Piece;
-import chess.domain.piece.Position;
+import chess.domain.position.Position;
 
 import java.util.Map;
