+++ src/main/java/chess/domain/board/ChessBoardFactory.java
@@ -3,6 +3,7 @@ package chess.domain.board;
 import chess.domain.camp.TeamColor;
 import chess.domain.chess.ChessGame;
 import chess.domain.piece.*;
+import chess.domain.position.Position;
 
 import java.util.HashMap;
 import java.util.Map;
