+++ src/main/java/chess/domain/piece/Direction.java
@@ -1,5 +1,7 @@
 package chess.domain.piece;
 
+import chess.domain.position.Position;
+
 import java.util.List;
 
 public enum Direction {
