+++ src/main/java/chess/domain/piece/Movable.java
@@ -1,5 +1,7 @@
 package chess.domain.piece;
 
+import chess.domain.position.Position;
+
 @FunctionalInterface
 public interface Movable {
     Move move = new Move();
