+++ src/test/java/chess/domain/board/ChessBoardTest.java
@@ -5,7 +5,7 @@ import chess.domain.chess.ChessGame;
 import chess.domain.piece.Pawn;
 import chess.domain.piece.Piece;
 import chess.domain.piece.PieceType;
-import chess.domain.piece.Position;
+import chess.domain.position.Position;
 import chess.domain.piece.Rook;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
