+++ src/test/java/chess/domain/position/PositionTest.java
@@ -1,5 +1,6 @@
-package chess.domain.piece;
+package chess.domain.position;
 
+import chess.domain.position.Position;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
